test: move things around to get passing
