+++ containers/ecr-viewer/src/app/tests/api/fhir-data.test.tsx
@@ -4,7 +4,7 @@
 import fs from "fs";
 import { S3Client, GetObjectCommand } from "@aws-sdk/client-s3";
 import { mockClient } from "aws-sdk-client-mock";
-import { GET } from "../api/fhir-data/route"; // Adjust the import path to your actual file path
+import { GET } from "../../api/fhir-data/route"; // Adjust the import path to your actual file path
 import { sdkStreamMixin } from "@smithy/util-stream";
 import { NextRequest } from "next/server";
 import { Readable } from "stream";
